Bump Netty to latest. Refactor existing container classes to use new API. Enhance SecurityContext to pass Netty ChannelHandlerContext, to allow finding remote IP etc.
